Rename package to com.labs1904.hwe
