+++ hbase-hello-world/src/test/scala/com/labs1904/hwe/MyAppTests.scala
@@ -1,4 +1,4 @@
-package com.kitmenke.hbase
+package com.labs1904.hwe
 
 import org.scalatest.FunSpec
