+++ scala-hello-world/src/main/scala/com/labs1904/hwe/exercises/Miscellaneous.scala
@@ -1,4 +1,4 @@
-package com.farrellw.hwe.exercises
+package com.labs1904.hwe.exercises
 
 object Miscellaneous {
   /*
