+++ scala-hello-world/src/main/scala/com/labs1904/hwe/exercises/RomanNumeral.scala
@@ -1,4 +1,4 @@
-package com.farrellw.hwe.exercises
+package com.labs1904.hwe.exercises
 
 // NOTE - For simplicity sake, solve assuming old roman numeral style. 4 is IIII and NOT IV
 object RomanNumeral {
