+++ scala-hello-world/src/main/scala/com/labs1904/hwe/worksheets/2-Collections.sc
