+++ scala-hello-world/src/test/scala/com/labs1904/hwe/exercises/WordCountTests.scala
@@ -1,4 +1,4 @@
-package com.farrellw.hwe.exercises
+package com.labs1904.hwe.exercises
 
 import org.scalatest.FunSpec
