+++ scala-hello-world/src/test/scala/com/labs1904/hwe/practice/CollectionsTests.scala
@@ -1,4 +1,4 @@
-package com.farrellw.hwe.practice
+package com.labs1904.hwe.practice
 
 import org.scalatest.FunSpec
